FastAPI + SQL for client orders/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-dotenv
psycopg2-binary
openai